Improve ToS Agreement on Topic submission
Right now the ToS is not visible or clickable on smaller devices. The
width is too large, and the lack of a max height prevents the user from
scrolling to see the whole modal.

This PR is a hacky improvement to just ensure that one is able to scroll
down if the text isn't visable. It also dictates that modals should be a
percentage of the total, not a fixed width size.