Add SponsorGroup and render sponsors according to level
We are trying to expose our sponsorship levels to be a bit more visible
on the site. Right now the sponsors are all listed in alphabetical order
with no precedence for sponsor levels. We want to call out significant
sponsors first (for example if Zoro was our top sponsor), rather than
listing them last because of the alphabetical.

This introduces a new model called SponsorGroup to let us be flexible
about naming. We can introduce new groups or change orders with the
list_priority as need be, and the groups will only be displayed if
there's a corresponding Sponsor in the Group.