Decouple Bootstrap CSS from site css (#328)
* Decouple Bootstrap CSS from site css

__Problem__

Years ago we merged all of the bootstrap CSS into one big monstrosity of
a CSS file. In it we then modified core bootstrap configurations to make
our site look as absolutely amazing as it does now. This meant that
upgrading bootstrap would be like performing open-heart monster surgery.

__Solution__

No more! Perform one round of surgery to remove the chipy-specific CSS
from Bootstrap internals ahead of a possible upgrade from bootstrap.

We don't think this change is going to be perfect. There will be some
minor differences, but we did do some work to ensure the home page
didn't have significant style differences. Given our desire to get on
latest bootstrap and to redesign the site, it's worth breaking a few
eggs.

Co-authored-by: elmq0022 raymondberg sophm256 <best-team-ever@example.com>